[executor-benchmark] refactor: extract pipeline creation out
